GEN_AI_Tasks/Task 1/code/requirements.txt
feedparser==6.0.12
beautifulsoup4==4.14.2
google-genai==1.66.0
Pillow==11.3.0
moviepy==2.2.1
gTTS==2.5.4
